requirements.txt
pandas
numpy
ta
python-telegram-bot
pybit
pydantic
pyyaml
requests
schedule
mcp
google-generativeai
# Novas dependências v2
sqlalchemy
psycopg2-binary
streamlit>=1.40.0
streamlit-autorefresh
plotly
flask
